requirements.txt
Flask~=1.1.2
gunicorn==20.0.4
PyMySQL==0.9.3
SQLAlchemy~=1.3.8
ParlPy==2.4.0
flask-cors==3.0.10
mysql-connector-python~=8.0.23
requests~=2.25.1
pandas~=1.2.3
google-cloud-secret-manager==2.2.0
google-cloud-logging==2.3.0
google-cloud-storage==1.36.2
GoogleAppEngineCloudStorageClient==1.9.22.1